requirements.txt
# Core LLM & Orchestration
openai
langchain==0.3.22
langchain-core==0.3.75
langchain-community==0.3.20
langchain-openai
langgraph

# Memory / Checkpointing
redis
langgraph-checkpoint-redis

# The Official Google Agent-to-Agent Protocol SDK
a2a-sdk[http-server]==0.2.4
uvicorn

# Tools & Protocols
mcp
pydantic

# Observability & Environment
langfuse==3.14.5
gradio
python-dotenv
httpx